Added filters with rest api compatible responses

+++ app/Config/Routes.php
@@ -13,13 +13,13 @@ $routes->post('users/authentication/login', 'UserAuthenticationController::login
 $routes->post('users/authentication/logout', 'UserAuthenticationController::logout_user');
 $routes->post('users/authentication', 'UserAuthenticationController::create_token');
 
-$routes->get('users', 'UserController::index'/*, ['filter' => 'admin']*/);
-$routes->get('users/(:num)', 'UserController::show/$1'/*, ['filter' => 'auth']*/);
+$routes->get('users', 'UserController::index'/*, ['filter' => 'rest_admin']*/);
+$routes->get('users/(:num)', 'UserController::show/$1'/*, ['filter' => 'rest_auth']*/);
 $routes->post('users', 'UserController::create');
-$routes->put('users/(:num)', 'UserController::update/$1'/*, ['filter' => 'auth']*/);
-$routes->delete('users/(:num)', 'UserController::delete/$1'/*, ['filter' => 'auth']*/);
+$routes->put('users/(:num)', 'UserController::update/$1'/*, ['filter' => 'rest_auth']*/);
+$routes->delete('users/(:num)', 'UserController::delete/$1'/*, ['filter' => 'rest_admin']*/);
 
 $routes->get('sessions/bookings/(:num)', 'SessionBookingController::show/$1');
 $routes->get('sessions/bookings/(:num)/(:num)', 'SessionBookingController::get_by_range/$1/$2');
-$routes->delete('sessions/bookings/(:num)', 'SessionBookingController::delete/$1');
-$routes->post('sessions/bookings', 'SessionBookingController::create');
+$routes->delete('sessions/bookings/(:num)', 'SessionBookingController::delete/$1', ['filter' => 'rest_auth']);
+$routes->post('sessions/bookings', 'SessionBookingController::create', ['filter' => 'rest_auth']);
